pyproject.toml
[project]
name = "Producer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "tenacity>=9.1.4",
]
